#175 Refactor: TagService, BookmarkTagService 통합
