+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/controller/BookmarkTagController.java
@@ -2,8 +2,7 @@ package com.kakao.linknamu.bookmarktag.controller;
 
 import com.kakao.linknamu.bookmarktag.dto.CreateBookmarkTagRequestDto;
 import com.kakao.linknamu.bookmarktag.dto.DeleteBookmarkTagRequestDto;
-import com.kakao.linknamu.bookmarktag.service.BookmarkTagDeleteService;
-import com.kakao.linknamu.bookmarktag.service.BookmarkTagSaveService;
+import com.kakao.linknamu.bookmarktag.service.BookmarkTagService;
 import com.kakao.linknamu.core.security.CustomUserDetails;
 import com.kakao.linknamu.core.util.ApiUtils;
 import lombok.RequiredArgsConstructor;
@@ -15,16 +14,15 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/api/tag")
 public class BookmarkTagController {
-	private final BookmarkTagSaveService bookmarkTagSaveService;
-	private final BookmarkTagDeleteService bookmarkTagDeleteService;
+	private final BookmarkTagService bookmarkTagService;
 
 	@PostMapping("/create/{bookmarkId}")
 	public ResponseEntity<?> createBookmarkTag(
 		@PathVariable(value = "bookmarkId") Long bookmarkId,
 		@RequestBody CreateBookmarkTagRequestDto requestDto,
 		@AuthenticationPrincipal CustomUserDetails userDetails) {
-		// 북마크 태크 생성
-		bookmarkTagSaveService.create(requestDto, userDetails.getUser(), bookmarkId);
+		// 북마크 태그 생성
+		bookmarkTagService.create(requestDto, userDetails.getUser(), bookmarkId);
 		return ResponseEntity.ok(ApiUtils.success(null));
 	}
 
@@ -33,7 +31,7 @@ public class BookmarkTagController {
 		@RequestBody DeleteBookmarkTagRequestDto requestDto,
 		@AuthenticationPrincipal CustomUserDetails userDetails) {
 		// 북마크 태그 제거
-		bookmarkTagDeleteService.deleteByTagIdAndBookmarkId(requestDto, userDetails.getUser());
+		bookmarkTagService.deleteByTagIdAndBookmarkId(requestDto, userDetails.getUser());
 		return ResponseEntity.ok(ApiUtils.success(null));
 	}
 }
