+++ linknamu/src/main/java/com/kakao/linknamu/share/service/category/GetCategoryFromEncodedIdService.java
@@ -2,7 +2,7 @@ package com.kakao.linknamu.share.service.category;
 
 import com.kakao.linknamu.bookmark.entity.Bookmark;
 import com.kakao.linknamu.bookmark.service.BookmarkService;
-import com.kakao.linknamu.bookmarktag.service.BookmarkTagReadService;
+import com.kakao.linknamu.bookmarktag.service.BookmarkTagService;
 import com.kakao.linknamu.category.entity.Category;
 import com.kakao.linknamu.category.service.CategoryService;
 import com.kakao.linknamu.core.dto.PageInfoDto;
@@ -23,9 +23,9 @@ import java.util.List;
 @Service
 public class GetCategoryFromEncodedIdService {
 	private final CategoryService categoryService;
-	private final BookmarkTagReadService bookmarkTagReadService;
 	private final AesEncryption aesEncryption;
 	private final BookmarkService bookmarkService;
+	private final BookmarkTagService bookmarkTagService;
 
 	public GetCategoryFromLinkResponseDto getCategory(String encodedCategoryId, Pageable pageable) {
 
@@ -39,7 +39,7 @@ public class GetCategoryFromEncodedIdService {
 		List<List<Tag>> tagListList = new ArrayList<>();
 
 		for (Bookmark bookmark : bookmarkPage.getContent()) {
-			List<Tag> tags = bookmarkTagReadService.findTagByBookmarkId(bookmark.getBookmarkId());
+			List<Tag> tags = bookmarkTagService.findTagByBookmarkId(bookmark.getBookmarkId());
 			tagListList.add(tags);
 		}
